+++ module/Depot/src/Depot/Process/Validation/DepotValidationProcessAwareTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Depot\Process\Validation;
+
+trait DepotValidationProcessAwareTrait
+{
+    protected DepotValidationProcess $depotValidationProcess;
+
+    public function setDepotValidationProcess(DepotValidationProcess $depotValidationProcess): void
+    {
+        $this->depotValidationProcess = $depotValidationProcess;
+    }
+}